+++ public/pages/CreateTransform/components/DefineTransforms/DefineTransforms.tsx
@@ -21,6 +21,7 @@ import { FieldItem, GROUP_TYPES, TransformGroupItem } from "../../../../../model
 import { TransformService } from "../../../../services";
 import { getErrorMessage } from "../../../../utils/helpers";
 import { isNumericMapping } from "../../utils/helpers";
+import PreviewTransform from "../PreviewTransform";
 
 interface DefineTransformsProps {
   transformService: TransformService;
@@ -48,32 +49,6 @@ export default function DefineTransforms({
   previewTransform,
 }: DefineTransformsProps) {
   let columns: EuiDataGridColumn[] = [];
-
-  const [previewColumns, setPreviewColumns] = useState<EuiDataGridColumn[]>([]);
-
-  const updatePreviewColumns = (): void => {
-    // setPreviewColumns([{id: "ex-col"}]);
-    // setPreviewColumns([]);
-    if (previewTransform.length) {
-      let tempCol: EuiDataGridColumn[] = [];
-      for (const [key, value] of Object.entries(previewTransform[0])) {
-        tempCol.push({
-          id: key,
-          actions: {
-            showHide: false,
-            showMoveLeft: false,
-            showMoveRight: false,
-            showSortAsc: false,
-            showSortDesc: false,
-          },
-        });
-      }
-      setPreviewColumns(tempCol);
-      setVisiblePreviewColumns(() => tempCol.map(({ id }) => id).slice(0, 5));
-    }
-    //Debug use
-    console.log("Preview columns: " + JSON.stringify(previewColumns));
-  };
 
   fields.map((field: FieldItem) => {
     const isNumeric = isNumericMapping(field.type);
@@ -103,9 +78,6 @@ export default function DefineTransforms({
                 },
               });
               onGroupSelectionChange(groupSelection);
-              previewColumns.push({
-                id: targetFieldName,
-              });
             },
             size: "xs",
             color: isNumeric ? "text" : "subdued",
@@ -146,7 +118,6 @@ export default function DefineTransforms({
                 sum: { field: field.label },
               };
               onAggregationSelectionChange(aggSelection);
-              console.log(Object.entries(selectedAggregations));
             },
             size: "xs",
             color: "text",
@@ -229,12 +200,10 @@ export default function DefineTransforms({
 
   const [loading, setLoading] = useState<boolean>(true);
   const [pagination, setPagination] = useState({ pageIndex: 0, pageSize: 10 });
-  const [previewPagination, setPreviewPagination] = useState({ pageIndex: 0, pageSize: 10 });
   const [from, setFrom] = useState<number>(0);
   const [size, setSize] = useState<number>(10);
   const [sortingColumns, setSortingColumns] = useState([]);
   const [visibleColumns, setVisibleColumns] = useState(() => columns.map(({ id }) => id).slice(0, 5));
-  const [visiblePreviewColumns, setVisiblePreviewColumns] = useState(() => previewColumns.map(({ id }) => id).slice(0, 5));
   const [data, setData] = useState([]);
   const [dataCount, setDataCount] = useState<number>(0);
   const [groupSelection, setGroupSelection] = useState<TransformGroupItem[]>(selectedGroupField);
@@ -259,10 +228,6 @@ export default function DefineTransforms({
     fetchData();
   }, []);
 
-  React.useEffect(() => {
-    updatePreviewColumns();
-  }, [previewTransform]);
-
   const onChangeItemsPerPage = useCallback(
     (pageSize) => {
       setPagination((pagination) => ({
@@ -283,23 +248,6 @@ export default function DefineTransforms({
     [setPagination]
   );
 
-  const onChangePreviewPerPage = useCallback(
-    (pageSize) => {
-      setPreviewPagination((previewPagination) => ({
-        ...previewPagination,
-        pageSize,
-        pageIndex: 0,
-      }));
-    },
-    [setPreviewPagination]
-  );
-  const onChangePreviewPage = useCallback(
-    (pageIndex) => {
-      setPreviewPagination((previewPagination) => ({ ...previewPagination, pageIndex }));
-    },
-    [setPreviewPagination]
-  );
-
   const onSort = useCallback(
     (sortingColumns) => {
       setSortingColumns(sortingColumns);
@@ -316,13 +264,6 @@ export default function DefineTransforms({
         return data[rowIndex]._source[correspondingTextColumnId] ? data[rowIndex]._source[correspondingTextColumnId] : "-";
       }
       return data[rowIndex]._source[columnId] ? data[rowIndex]._source[columnId] : "-";
-    }
-    return "-";
-  };
-
-  const renderPreviewCellValue = ({ rowIndex, columnId }) => {
-    if (previewTransform.hasOwnProperty(rowIndex)) {
-      return previewTransform[rowIndex][columnId] ? previewTransform[rowIndex][columnId] : "-";
     }
     return "-";
   };
@@ -387,26 +328,7 @@ export default function DefineTransforms({
       </EuiText>
       <EuiSpacer size="s" />
       {previewTransform.length ? (
-        <EuiDataGrid
-          aria-label="Preview transforms"
-          columns={previewColumns}
-          columnVisibility={{ visibleColumns: visiblePreviewColumns, setVisibleColumns: setVisiblePreviewColumns }}
-          rowCount={previewTransform.length}
-          renderCellValue={renderPreviewCellValue}
-          // sorting={{ columns: sortingColumns, onSort }}
-          pagination={{
-            ...previewPagination,
-            pageSizeOptions: [5, 10, 20, 50],
-            onChangeItemsPerPage: onChangePreviewPerPage,
-            onChangePage: onChangePreviewPage,
-          }}
-          toolbarVisibility={{
-            showColumnSelector: true,
-            showStyleSelector: false,
-            showSortSelector: false,
-            showFullScreenSelector: false,
-          }}
-        />
+        <PreviewTransform previewTransform={previewTransform} />
       ) : (
         <EuiPanel>
           <EuiEmptyPrompt
